+++ src/main/java/rocks/zipcode/atm/bank/AccountData.java
@@ -9,10 +9,10 @@ public final class AccountData {
     private final String name;
     private final String email;
 
-    private final int balance;
+    private final float balance;
     private  String alert;
 
-    AccountData(int id, String name, String email, int balance) {
+    AccountData(int id, String name, String email, float balance) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -39,7 +39,7 @@ public final class AccountData {
 
     public String getAlert() { return alert;}
 
-    public int getBalance() {
+    public float getBalance() {
         if(this.balance > 0)
         {// reset alert
             alert = "";
